♻️Refactor : 홈 리펙토링
